# laravel9-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto V_Lab",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories/{user_id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Listar categorias",
                "description": "Listar todas as categorias de um usuário",
                "operationId": "722185be9ae5a3dfbd2a26cbb207091b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de categorias do usuário"
                    },
                    "404": {
                        "description": "Nenhuma categoria encontrada para o usuário"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Criar categoria",
                "description": "Criar uma nova categoria para o usuário",
                "operationId": "580461dce4d15406e77ad6e2c4f408b9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nova Categoria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria cadastrada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação dos dados"
                    },
                    "409": {
                        "description": "Categoria já cadastrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/categories/{user_id}/delete/{category_id}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Apagar categoria e suas transações associadas",
                "description": "Apaga uma categoria e todas as transações associadas a ela.",
                "operationId": "2b8fdf19edab8558a02d19b74dedb4c5",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria e transações associadas deletadas com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar Usuários",
                "description": "Endpoint para listar todos os usuários cadastrados.",
                "operationId": "f43dfbe89a69790f581f1f8da36054be",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cadastrar Usuário",
                "description": "Endpoint para cadastrar um novo usuário.",
                "operationId": "17072b1d1ca1d104af6eb3a27746169b",
                "requestBody": {
                    "description": "Dados do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuário",
                                        "type": "string",
                                        "example": "Vinicius Fialho"
                                    },
                                    "cpf": {
                                        "description": "CPF do usuário",
                                        "type": "string",
                                        "example": "10607446439"
                                    },
                                    "email": {
                                        "description": "Endereço de e-mail do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "example": "vaf2@cin.ufpe.br"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/{id}/edit": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar Usuário",
                "description": "Endpoint para editar um usuário existente.",
                "operationId": "b8036f23f8b59de6fe43998b325930ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser editado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do usuário a serem atualizados",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "E-mail do usuário",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF do usuário",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário editado com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Erro de validação dos dados de entrada",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}